src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/base-in-shadow-tree-expected.txt

CONSOLE ERROR: line 13: Uncaught TypeError: container.createShadowRoot is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

tehepero
tehepero
